\name{MED-package}
\alias{MED-package}
\docType{package}
\title{
Nonparametric estimation of natural mediation effects by tilted balancing.
}
\description{
This package provides a user-friendly interface for nonparametric efficient inference of natural mediation effects effects for observational data. The package provides point estimates for average treatment effects, natural direct effects and natural indirect effects.  The package also allows inference by consistent variance estimators.
}
\details{

\tabular{ll}{
Package: \tab MED\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-02-09\cr
License: \tab GPL (>= 2) \cr
}

The package includes the following functions:

\tabular{ll}{
\code{\link{MED}}: \tab Estimate the Natural Indirect and Direct Effects \cr
\code{\link{summary.MED}}: \tab \code{summary} method for class \code{"MED"} \cr
}
}

\author{
Gary Chan, based on the ATE package developed together with Asad Haris.

Maintainer: Gary Chan <kcgchan@uw.edu>
}

\references{
Chan, K. C. G., Imai, K, Yam, S. C. P. and Zhang, Z. (2016). "Efficient Nonpmarametric Estimation of Causal Mediation Effects.", under review.
}

\keyword{ package }
\seealso{
\code{\link{MED}}
}
