% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meDreD.R
\name{meDreD}
\alias{meDreD}
\title{Estimation of the mediated moderation (meD) and remaining moderation (reD)}
\usage{
meDreD(
  data_in,
  Mnames,
  Cnames,
  nuisance_estimation = "mlr",
  Yfamily = "binomial",
  estimator = "onestep",
  num_folds = 5
)
}
\arguments{
\item{data_in}{a \code{data.frame} containing the observed data.
In "data_in", column "tt" is the treatment assignment ("tt" is coded as 0 for individuals in the control condition and as 1 for individuals in the treatment condition);
column "Y" is the outcome.}

\item{Mnames}{a character vector of the names of the columns in "data_in" that correspond to mediators (M).}

\item{Cnames}{a character vector of the names of the columns in "data_in" that correspond to baseline covariates (C).}

\item{nuisance_estimation}{a character specifying the method for estimating the nuisance models. Currently supported options include: "linear" and "mlr". For "linear", the nuisance models are estimated using generalized linear regressions with linear terms of the covariates. For "mlr", the nuisance models are estimated using the "SuperLearner" R package (<https://cran.r-project.org/package=SuperLearner>) with an ensemble of machine learning methods (the specific methods can be modified).}

\item{Yfamily}{a character specifying the link function for a generalized linear model for the outcome. Currently supported options include "gaussian" and "binomial".}

\item{estimator}{a character specifying the estimator for the mediated moderation (meD) and remaining moderation (reD). Currently supported options include "onestep" and "tml", denoting the one-step estimator and targeted minimum loss estimator, respectively.}

\item{num_folds}{the number of folds used for the cross-fitting procedure.}
}
\description{
Estimation of the mediated moderation (meD) and remaining moderation (reD)
}
\examples{
 # data(data_in)
 # data_in <- read.csv("data/data_in.csv", header = TRUE)
 # data_in$obs_weights_Y <- rep(1, nrow(data_in))
 # Mnames <- grep("Mdat", colnames(data_in), value = T)
 # Cnames <- grep("Cdat", colnames(data_in), value = T)
 # out <- meDreD(
 # data_in = data_in,
 # Mnames = Mnames,
 # Cnames = Cnames,
 # nuisance_estimation = "linear",
 # Yfamily = "binomial",
 # estimator = "onestep",
 # num_folds = 5)


}
